sso profile service and tests
